requirements.txt
torch>=1.0.0
torchvision>=0.4.2
tqdm>=4.31.1
numpy
pillow>=4.3.0
git+https://github.com/jonbarron/robust_loss_pytorch
gpustat>=0.6.0